TemperatureConverter/app/src/test/java/com/vogella/android/temperatureconverter/ConverterUtilTest.java
package com.vogella.android.temperatureconverter;

/**
 * Created by amritachowdhury on 7/28/16.
 */

import android.graphics.Bitmap;

import static org.junit.Assert.*;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.RobolectricGradleTestRunner;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.annotation.Config;

@RunWith(RobolectricGradleTestRunner.class)
@Config(constants=BuildConfig.class, sdk = 18)
public class ConverterUtilTest {

    @Test
    public void testConvertFahrenheitToCelsius() {
        float actual = ConverterUtil.convertCelsiusToFahrenheit(100);
        // expected value is 212
        float expected = 212;
        // use this method because float is not precise
        assertEquals("Conversion from celsius to fahrenheit failed", expected,
                actual, 0.001);
    }

    @Test
    public void testConvertCelsiusToFahrenheit() {
        float actual = ConverterUtil.convertFahrenheitToCelsius(212);
        // expected value is 100
        float expected = 100;
        // use this method because float is not precise
        assertEquals("Conversion from celsius to fahrenheit failed", expected,
                actual, 0.001);
    }

    @Test
    public void testBitmapScaling() {
        Bitmap bitmap = Bitmap.createBitmap(200, 200, Bitmap.Config.ARGB_8888);
        Bitmap actual= ConverterUtil.generateFullBleedBitmap(bitmap, 2500);
        assertEquals("Bitmap width is incorrect", 2500, actual.getWidth());
        assertEquals("Bitmap height is incorrect", 2500, actual.getHeight());
    }
}
